flag: define SetHelpOutput and HelpOutput to specify the output destination when help flag is explicitly specified
When the help flag is explicitly specified, I want to put usage to Stdout instead of Stderr, because I want to read the usage using pager command such as "less" by pipe the output.

With this change, we can specify the output destination when the help flag is explicitly specified with fs.SetHelpOutput method.

And the destination can be retrieved by fs.HelpOutput. This method is useful when defining the custom fs.Usage.

The HelpOutput method returns the destination specified in the SetHelpOutput method only when the help flag is explicitly specified. It means, if usage is output due to a flag parsing error, the message will be output to the destination of fs.Output as before.